assistant/requirements.txt
# ========== LangChain 核心 ==========
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.3.19

# ========== 前端界面 ==========
streamlit==1.41.1

# ========== 文档处理 ==========
unstructured[all-local]==0.16.12

# ========== Token 计数 ==========
tiktoken==0.8.0

# ========== 环境变量管理 ==========
python-dotenv==1.0.0